disable attachment button when saving
