+++ lib/pages/transaction.dart
@@ -1,7 +1,6 @@
 import 'dart:async';
 
 import 'package:async/async.dart';
-import 'package:badges/badges.dart' as badges;
 import 'package:chopper/chopper.dart' show Response;
 import 'package:flutter/services.dart';
 import 'package:flutter_sharing_intent/model/sharing_file.dart';
@@ -25,7 +24,6 @@ import 'package:waterflyiii/settings.dart';
 import 'package:waterflyiii/stock.dart';
 import 'package:waterflyiii/timezonehandler.dart';
 import 'package:waterflyiii/widgets/autocompletetext.dart';
-import 'package:waterflyiii/widgets/materialiconbutton.dart';
 
 final Logger log = Logger("Pages.Transaction");
 
@@ -1071,70 +1069,6 @@ class _TransactionBudgetState extends State<TransactionBudget> {
           ),
         ),
       ],
-    );
-  }
-}
-
-class AttachmentButton extends StatefulWidget {
-  final List<AttachmentRead>? attachments;
-  final Future<void> Function() onPressed;
-
-  const AttachmentButton({
-    super.key,
-    required this.attachments,
-    required this.onPressed,
-  });
-
-  @override
-  State<AttachmentButton> createState() => _AttachmentButtonState();
-}
-
-class _AttachmentButtonState extends State<AttachmentButton> {
-  late bool _hasAttachments;
-
-  final Logger log = Logger("Pages.Transaction.AttachmentButton");
-
-  @override
-  void initState() {
-    super.initState();
-
-    log.finest(() => "initState()");
-    _hasAttachments = widget.attachments?.isNotEmpty ?? false;
-  }
-
-  @override
-  void didUpdateWidget(covariant AttachmentButton oldWidget) {
-    super.didUpdateWidget(oldWidget);
-
-    log.finest(() => "didUpdateWidget()");
-    _hasAttachments = widget.attachments?.isNotEmpty ?? false;
-    log.finest(() => "_hasAttachments: $_hasAttachments");
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    log.finest(() => "build(${widget.attachments?.length ?? "null"})");
-
-    return badges.Badge(
-      badgeContent: Text(
-        widget.attachments?.length.toString() ?? "..",
-        style: Theme.of(context).textTheme.labelMedium!.copyWith(
-          color: Theme.of(context).colorScheme.onSurfaceVariant,
-        ),
-      ),
-      showBadge: _hasAttachments,
-      badgeStyle: badges.BadgeStyle(
-        badgeColor: Theme.of(context).colorScheme.surfaceContainerHighest,
-      ),
-      badgeAnimation: const badges.BadgeAnimation.scale(
-        animationDuration: animDurationEmphasized,
-        curve: animCurveEmphasized,
-      ),
-      child: MaterialIconButton(
-        icon: Icons.attach_file,
-        tooltip: S.of(context).transactionAttachments,
-        onPressed: _savingInProgress ? null : widget.onPressed,
-      ),
     );
   }
 }
